backend/requirements.txt
Django
psycopg2
django-cors-headers
django-guardian
djangorestframework
djangorestframework-simplejwt
requests
Pillow
sendgrid
sendgrid-mail
google-cloud-storage